Add deprecation warning for Tilt.prefer
This used to have a deprecation warning, but it was removed about
a decade ago, for greater compatibility with code that wanted to
support both Tilt 1.x and 2.x.  Tilt 1.x compatibility should no
longer be a concern currently, so officially deprecate this with
a warning, for planned removal in Tilt 2.3.