Add Tilt::StaticTemplate for templates that return the same output for every render
A majority of the template libraries Tilt supports are static templates, where
the returned value is the same every time the template is rendered.  Previously,
Tilt used a normal template subclass for these templates, resulting in a lot of
redundancy.  This combines all of the shared behavior into a single StaticTemplate
subclass, and provides a StaticTemplate.subclass method to make it easy to define
static template subclasses. This method instance_execs the block given in the
StaticTemplate instance, and the block should return the rendered data.

One behavior change here is that Tilt will attempt to render static templates when
the template is created, to cache the output, even if it is never rendered. That
is unlikely to cause a problem in practice, because there is no point in creating
a template and not using it.