Make SassTemplate ignore unsupported options when using sass-embedded
When using sass and sassc, unsupported options are ignored, and
that should be the same behavior when using sass-embedded.

Use #parameters to attempt to determine the supported options. I
didn't look at all versions of sass-embedded to determine if the
code to get the supported options will work in all cases, so if
there are any problems determining the supported options, treat
all options as supported (same as previous behavior).

Add a test that uses an unsupported option.