Support fixed locals for templates
Tilt's default approach of allowing any locals to be templates is quite
flexible, but it has a number of issues:

* It is inefficient, especially for large templates that are called with
  many combinations of locals.
* It hides issues if unused local variable names are passed to the template
* It does not support default values for local variables
* It does not support required values for local variables
* It does not support cases where you want to pass values via a keyword splat
* It does not support named blocks

Fixed locals can fix all of these issues.  With fixed locals, instead of the
compiled template methods taking a single positional argument, which is
passed as a hash, they can take arbitrary arguments provided by the user.
Commonly, the arguments will be required and positional keyword arguments,
but this also supports single positional argument without keywords, or using
a keyword splat with or without other keywords (useful when passing keywords
provided to the template to other methods), as well as used a named block
(useful when passing the block to other methods).

The expected way to use fixed locals is for them to be embedded in templates,
but that could cause backwards compatibility issues, so that needs to be
manually enabled by calling:

```ruby
Tilt.extract_fixed_locals = true
```

There are three supported ways of specifying fixed locals, in order of
preference:

* :fixed_locals template option
* embedded fixed locals magic comment (if Tilt.extract_fixed_locals = true)
* :default_fixed_locals template option

This is tested with the erubi and string templates, but it should work for
all templates that support locals.  For better backwards compatibility
with external templates, this uses a frozen empty array as the locals_keys
argument passed to internal methods, when fixed locals are used.

Libraries that use tilt and call compiled_method directly to get the compiled
method will probably need to be updated to call the returned compiled methods
with a keyword splat of the locals hash instead of passing the locals hash
directly as a positional argument, if the template supports fixed_locals
(which you can test by calling fixed_locals? on the template object).