Drop JRuby 9.1-9.3 CI
JRuby 9.1 and 9.2 are no longer supported on ubuntu-master.
JRuby 9.3 apparently has issues with keyword arguments that
result in ArgumentError not being raised for method calls that
pass keywords, when the method does not accept any arguments,
which breaks the fixed locals support.

JRuby 9.1 and 9.2 are definitely unsupported.  I'm not sure if
JRuby 9.3 is unsupported, but even if it is, it won't be for
much longer.

I'm open to readding these to CI, if someone else is willing to
do the work to get CI passing with them.